requirements.txt
# ============================================================
# Fraud Detection API — Dependencies
# ============================================================

# --- Data & ML ---
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
geopy>=2.4

# --- Kaggle data download ---
kagglehub>=0.3
opendatasets>=0.1.22

# --- API ---
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0

# --- Serialization ---
joblib>=1.3

# --- Logging & Config ---
python-dotenv>=1.0
loguru>=0.7

# --- Testing ---
pytest>=8.0
httpx>=0.27

# --- Dev tools (optional) ---
black>=24.0
isort>=5.13
